# docs/swagger.yaml
basePath: /api
definitions:
  model.LoginReq:
    properties:
      login:
        type: string
      pass:
        type: string
    type: object
  model.LoginResp:
    properties:
      access_token:
        type: string
      expires_in:
        type: integer
      token_type:
        type: string
    type: object
  model.RegisterReq:
    properties:
      login:
        type: string
      pass:
        type: string
    type: object
  model.RegisterResp:
    properties:
      message:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: bitop@spatecon.ru
    name: API Support
    url: https://vk.com/bmstu_schedule
  description: API для работы с экзопланетами и звездами
  license:
    name: AS IS (NO WARRANTY)
  title: ExoCalc API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Аутентификация пользователя с получением JWT токена
      parameters:
      - description: Данные для входа
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/model.LoginReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.LoginResp'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      summary: Авторизация пользователя
      tags:
      - Auth
  /auth/logout:
    post:
      consumes:
      - application/json
      description: Завершение сессии пользователя и инвалидация токена
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Выход пользователя
      tags:
      - Auth
...
  /auth/register:
    post:
      consumes:
      - application/json
      description: Создание нового пользователя с ролью Client
      parameters:
      - description: Данные для регистрации
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/model.RegisterReq'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.RegisterResp'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      summary: Регистрация нового пользователя
      tags:
      - Auth
  /auth/register-astronomer:
    post:
      consumes:
      - application/json
      description: Создание нового пользователя с ролью Astronomer
      parameters:
      - description: Данные для регистрации
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/model.RegisterReq'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.RegisterResp'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      summary: Регистрация астронома
      tags:
      - Auth
...
  /calculate-exoplanets/updateStarComment:
    put:
      consumes:
      - application/json
      description: Обновляет комментарий для конкретной звезды в заявке
      parameters:
      - description: ID заявки
        in: query
        name: selected_stars_id
        required: true
        type: integer
      - description: ID звезды
        in: query
        name: star_id
        required: true
        type: integer
      - description: Комментарий
        in: body
        name: input
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Обновление комментария для расчетов экзопланет
      tags:
      - CalculateExoplanets
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: 'Введите токен в формате: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"